refactor: move validators to utils
