+++ seu_lourival/lib/core/utils/string_regex.dart
@@ -1,0 +1,7 @@
+extension StringRegex on String {
+  bool isValidName() {
+    return RegExp(
+            r'^[A-ZÀ-ŸA-zÀ-ÿ][A-ZÀ-ŸA-zÀ-ÿ]+\s([A-ZÀ-ŸA-zÀ-ÿ]\s?)*[A-ZÀ-ŸA-zÀ-ÿ][A-ZÀ-ŸA-zÀ-ÿ]+$')
+        .hasMatch(this);
+  }
+}
